% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_game.R
\name{create_game}
\alias{create_game}
\title{Créer une nouvelle partie de Slitherlink}
\usage{
create_game(n, difficulte = "Facile","Moyen", "Difficile", "Expert")
}
\arguments{
\item{n}{taille de la grille}
}
\value{
un objet de type slitherlink
}
\description{
Cette fonction crée une grille vide de Slitherlink de taille n x n.
}
