% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grille.R
\name{dessiner_grille}
\alias{dessiner_grille}
\title{Dessiner la grille avec points agrandis}
\usage{
dessiner_grille(niveau, segments_h, segments_v, afficher_erreurs = FALSE)
}
\description{
Dessiner la grille avec points agrandis
}
