% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clues.R
\name{get_clues}
\alias{get_clues}
\title{Générer une grille de Slitherlink}
\usage{
get_clues(n, difficulte = "Facile")
}
\arguments{
\item{n}{taille de la grille}
}
\value{
matrice des indices
}
\description{
Cette fonction génère ou extrait les indices nécessaires au jeu Slitherlink.
}
