% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{is_solved}
\alias{is_solved}
\title{Verifier si un puzzle Slitherlink est resolu}
\usage{
is_solved(puzzle)
}
\arguments{
\item{puzzle}{Un objet de classe \code{slitherlink}.}
}
\value{
Une liste contenant :
\describe{
\item{solved}{\code{TRUE} si les trois conditions sont remplies.}
\item{clues_ok}{\code{TRUE} si tous les indices sont respectes.}
\item{vertices_ok}{\code{TRUE} si tous les points ont 0 ou 2 aretes.}
\item{single_loop}{\code{TRUE} si les aretes forment une seule boucle.}
\item{bad_clues}{Matrice booleenne aux cases dont l'indice est viole.}
\item{bad_vertices}{Matrice booleenne aux points fautifs.}
\item{n_loops}{Nombre de composantes connexes d'aretes.}
}
}
\description{
Verifier les trois regles du Slitherlink :
\enumerate{
\item Chaque indice de case est respecte (nombre d'aretes tracees
autour de la case = indice).
\item A chaque point, il y a exactement 0 ou 2 aretes incidentes.
\item Les aretes tracees forment une seule boucle fermee.
}
}
\examples{
clues <- matrix(c(NA, 3, NA, 2,
                  2, NA, 1, NA,
                  NA, 0, NA, 3), nrow = 3, byrow = TRUE)
p <- new_puzzle(clues)
is_solved(p)

}
