% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzle.R
\name{new_puzzle}
\alias{new_puzzle}
\title{Creer un puzzle Slitherlink}
\usage{
new_puzzle(clues)
}
\arguments{
\item{clues}{Une matrice \verb{n x m} d'entiers entre 0 et 3, avec \code{NA}
pour les cases sans indice.}
}
\value{
Un objet de classe \code{slitherlink}, qui est une liste contenant :
\describe{
\item{clues}{La matrice d'indices.}
\item{h_edges}{Matrice \verb{(n+1) x m} des aretes horizontales (0/1).}
\item{v_edges}{Matrice \verb{n x (m+1)} des aretes verticales (0/1).}
\item{n}{Nombre de lignes de cases.}
\item{m}{Nombre de colonnes de cases.}
}
}
\description{
Construit un objet de classe \code{slitherlink} a partir d'une matrice
d'indices. Les aretes (horizontales et verticales) sont initialisees
a 0 (non tracees).
}
\examples{
clues <- matrix(c(NA, 3, NA, 2,
                  2, NA, 1, NA,
                  NA, 0, NA, 3), nrow = 3, byrow = TRUE)
p <- new_puzzle(clues)
p

}
