% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.slitherlink}
\alias{plot.slitherlink}
\title{Afficher un puzzle Slitherlink}
\usage{
\method{plot}{slitherlink}(x, show_grid = TRUE, ...)
}
\arguments{
\item{x}{Un objet de classe \code{slitherlink}.}

\item{show_grid}{Logique. Si \code{TRUE}, affiche les aretes non tracees en
gris clair pour visualiser la grille. Defaut : \code{TRUE}.}

\item{...}{Arguments supplementaires (non utilises).}
}
\value{
Un objet \code{ggplot}.
}
\description{
Dessine un puzzle Slitherlink avec \code{ggplot2} : les points de la grille,
les indices dans les cases, et les aretes actuellement tracees.
}
\examples{
clues <- matrix(c(NA, 3, NA, 2,
                  2, NA, 1, NA,
                  NA, 0, NA, 3), nrow = 3, byrow = TRUE)
p <- new_puzzle(clues)
plot(p)

}
