% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{puzzle_easy}
\alias{puzzle_easy}
\title{Puzzle Slitherlink facile (3 x 4)}
\format{
Un objet de classe \code{slitherlink}.
}
\usage{
puzzle_easy
}
\description{
Une petite grille Slitherlink de 3 lignes sur 4 colonnes,
utile pour tester l'application ou se familiariser avec le jeu.
}
\keyword{datasets}
