% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{puzzle_hard}
\alias{puzzle_hard}
\title{Puzzle Slitherlink difficile (7 x 7)}
\format{
Un objet de classe \code{slitherlink}.
}
\usage{
puzzle_hard
}
\description{
Une grande grille pour les amateurs de defi.
}
\keyword{datasets}
