% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{puzzle_medium}
\alias{puzzle_medium}
\title{Puzzle Slitherlink moyen (5 x 5)}
\format{
Un objet de classe \code{slitherlink}.
}
\usage{
puzzle_medium
}
\description{
Une grille de taille moyenne basee sur l'exemple du sujet.
}
\keyword{datasets}
