% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzle.R
\name{toggle_edge}
\alias{toggle_edge}
\title{Basculer l'etat d'une arete (tracee / non tracee)}
\usage{
toggle_edge(p, type, r, c)
}
\arguments{
\item{p}{Un objet de classe \code{slitherlink}.}

\item{type}{Caractere : \code{"h"} pour horizontale, \code{"v"} pour verticale.}

\item{r}{Indice de ligne de l'arete dans la matrice correspondante.}

\item{c}{Indice de colonne de l'arete dans la matrice correspondante.}
}
\value{
L'objet \code{slitherlink} modifie.
}
\description{
Basculer l'etat d'une arete (tracee / non tracee)
}
