% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grille.R
\name{verifier_boucle_unique}
\alias{verifier_boucle_unique}
\title{Vérifier s'il n'y a qu'une seule boucle (BFS)}
\usage{
verifier_boucle_unique(segments_h, segments_v)
}
\description{
Vérifier s'il n'y a qu'une seule boucle (BFS)
}
