% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_solved.R
\name{is_solved}
\alias{is_solved}
\title{Vérifier si la solution est correcte}
\usage{
is_solved(game)
}
\arguments{
\item{game}{objet slitherlink}
}
\value{
TRUE ou FALSE
}
\description{
Cette fonction vérifie si la solution du Slitherlink est correcte.
}
