% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzle.R
\name{toggle_edge}
\alias{toggle_edge}
\title{Basculer l'etat d'une arete}
\usage{
toggle_edge(p, type, r, c)

toggle_edge(p, type, r, c)
}
\arguments{
\item{p}{Un objet de classe \code{slitherlink}.}

\item{type}{Caractere, "h" pour une arete horizontale, "v" pour une verticale.}

\item{r, c}{Coordonnees de l'arete}

\item{c}{Indice de la colonne.}

\item{r}{Indice de la ligne.}
}
\value{
L'objet \code{slitherlink} avec l'arete modifiee.
}
\description{
Basculer l'etat d'une arete

Basculer l'etat d'une arete
}
